[move] BufferScopeの定義場所を移動
